Implments Gate facades

+++ app/Http/Controllers/QuestionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\NewQuestionRequest;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -76,7 +77,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit',compact('question'));
+        if(Gate::allows('update-question', $question)) {
+            return view('questions.edit',compact('question'));
+        }
+        abort(403,"You do not have access to given action");
     }
 
     /**
@@ -100,6 +104,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(Gate::denies('delete-question', $question)) {
+            abort(403,"No permission on given action");
+        }
         $question->delete();
 
         return redirect('/questions')->with('success',"The question has been removed!");
